requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.7
pluggy==0.13.1
protobuf==3.14.0
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyodbc==4.0.30
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-access==1.0.8
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
